% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{Clinical ordinal endpoints and treatments assignment for 200 patient}
\format{
A data frame with 200 rows and 2 variables:
\describe{
  \item{response}{outcome}
  \item{treatment}{0 denotes control, 1 denotes treatment}
  ...
}
}
\usage{
example.data
}
\description{
A dataset containing the ordinal outcomes and corresponding groups.
}
\keyword{datasets}
